requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
environs==11.0.0
gunicorn==22.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
